third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-weight-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['font-weight'] = "calc(100 + (sign(20cqw - 10px) * 5))" should set the property value
  assert_equals: serialization should be canonical expected "calc(100 + (5 * sign(20cqw - 10px)))" but got "calc(100 + sign(20cqw - 10px) * 5)"
Harness: the test ran to completion.
